hello world i changed many things thanks

+++ Hotel/Hotel/src/hotel/classes/ExtraOrders.java
@@ -56,7 +56,7 @@ public class ExtraOrders {
     
     public int calculateTotal()
     {
-       return item.getPrice() * quantity;
+       return (int) (item.getPrice() * quantity);
     }
